eventDetails data collection
Implementationof data collection for all event details... Right now the new version is able to insert to the Event Points table and also collect all the stats that are needed for League Stats. Full implementation is still waiting on Database changes which need to be discussed. (Also implementation right now is not generic, thats the next step)