eventDetails generalization work in progress
Generalized most of eventDetails, to the point where it inserts all needed information into the EventPoints table... Working on the strange error that is occuring on get requests at the moment. Once that blockage is fixed then we will have everything needed to insert into League Stats. Also, added a few node modules to deal with BIGINT and JSON parsing to BIGINT.